update azure storage account entity
